requirements.txt
streamlit
google-generativeai
openai
Pillow
requests
beautifulsoup4
youtube-transcript-api
